migrate IsPluginCommand to plugins pkg
