+++ pkg/cmd/root.go
@@ -61,7 +61,7 @@ var rootCmd = &cobra.Command{
 		telemetryMetadata.SetUserAgent(useragent.GetEncodedUserAgent())
 
 		// plugins send their own telemetry due to having richer context than the CLI does
-		if !stripe.IsPluginCommand(cmd) {
+		if !plugins.IsPluginCommand(cmd) {
 			// record command invocation
 			sendCommandInvocationEvent(cmd.Context())
 		}
