+++ pkg/stripe/analytics_telemetry.go
@@ -231,17 +231,3 @@ func TelemetryOptedOut(optoutVar string) bool {
 
 	return optoutVar == "1" || optoutVar == "true"
 }
-
-// IsPluginCommand returns true if the command invoked is for a plugin
-// flase otherwise
-func IsPluginCommand(cmd *cobra.Command) bool {
-	isPlugin := false
-
-	for key, value := range cmd.Annotations {
-		if key == "scope" && value == "plugin" {
-			isPlugin = true
-		}
-	}
-
-	return isPlugin
-}
